refactor: reorganize services into api
