+++ client/src/pages/Agentes.tsx
@@ -43,7 +43,7 @@ import {
 } from '@/types/agent';
 
 // Services
-import { agentService } from '@/services/agentService';
+import { agentService } from '@/api/agentService';
 
 // Utils
 import { cn } from '@/lib/utils';
